med-questionnaire/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
django-cors-headers>=4.0
scikit-learn>=1.3
numpy>=1.24
joblib>=1.3
